Get all tests passing, revert reorganization
