+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
-- 
+- Redesign component access checking. Complex queries will now pass the component access checker in more cases.
+- Improve diagnostics for component access errors.
+- Redesign `Config` interface to hide implementation details. (Now named `HandlerConfig`).
+- Removed `bit_set` module from public API.
 
 ## 0.4.2 - 2024-03-24
