+++ src/lib.rs
@@ -50,7 +50,7 @@ pub mod prelude {
     pub use crate::event::{
         Despawn, Event, EventId, EventMut, Insert, Receiver, ReceiverMut, Remove, Sender, Spawn,
     };
-    pub use crate::fetch::{Fetcher, FetchError, Single, SingleError, TrySingle};
+    pub use crate::fetch::{Fetcher, GetError, Single, SingleError, TrySingle};
     pub use crate::handler::{Handler, HandlerId, HandlerParam, IntoHandler};
     pub use crate::query::{Has, Not, Or, Query, ReadOnlyQuery, With, Xor};
     pub use crate::world::World;
